% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip.R
\name{bert_type}
\alias{bert_type}
\title{Pre-Trained BERT Model}
\usage{
bert_type(values = torchtransformers::available_berts())
}
\arguments{
\item{values}{A character vector indicating the names of available models.
The default includes 31 possible values, so we recommend you select
specific models that are likely to work on your hardware.}
}
\value{
A parameter that can be tuned with the \code{tune} package.
}
\description{
The pre-trained BERT model that will be fine-tuned for a model.
}
\examples{
if (rlang::is_installed("dials")) {
  bert_type()
}
}
