% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert_classification-fit.R
\name{.bert_classification_bridge}
\alias{.bert_classification_bridge}
\title{Bridge between hardhat and Implementation}
\usage{
.bert_classification_bridge(
  processed,
  model_name = "bert_tiny_uncased",
  n_tokens = torchtransformers::config_bert(model_name, "max_tokens"),
  ...
)
}
\arguments{
\item{processed}{Processed inputs molded by hardhat.}

\item{model_name}{Character; which flavor of BERT to use. See
\code{\link[torchtransformers]{available_berts}} for known models.}

\item{...}{Currently unused but included to expand into more luz options.}
}
\value{
A bert_classification model object.
}
\description{
Bridge between hardhat and Implementation
}
\keyword{internal}
