% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert_classification-fit.R
\name{.bert_classification_impl}
\alias{.bert_classification_impl}
\title{Create and Fit the Model}
\usage{
.bert_classification_impl(
  predictors,
  outcome,
  model_name = "bert_tiny_uncased",
  n_tokens = torchtransformers::config_bert(model_name, "max_tokens"),
  ...
)
}
\arguments{
\item{predictors}{A tibble containing one or two character columns.}

\item{outcome}{A factor of output classes associated with the predictors.}

\item{model_name}{Character; which flavor of BERT to use. See
\code{\link[torchtransformers]{available_berts}} for known models.}

\item{n_tokens}{Integer scalar; the number of tokens expected for each
example.}

\item{...}{Currently unused but included to expand into more luz options.}
}
\value{
The fitted model with class \code{luz_module_fitted}.
}
\description{
Create and Fit the Model
}
\keyword{internal}
