% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{.check_luz_model_serialization}
\alias{.check_luz_model_serialization}
\title{Check Pointers in a Luz Model}
\usage{
.check_luz_model_serialization(luz_model)
}
\arguments{
\item{luz_model}{The model to check.}
}
\value{
The model, invisibly.
}
\description{
The \code{.check$ptr} tensor in a \code{bert_linear} model allows us to check whether
the model contains invalid pointers. If it does, restore the
\code{serialized_state_dict} into the model (in-place).
}
\keyword{internal}
