% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{.check_predictors_are_character}
\alias{.check_predictors_are_character}
\title{Identify Non-Character Predictors}
\usage{
.check_predictors_are_character(predictors)
}
\arguments{
\item{predictors}{"Data-like" predictors, in this case always a tibble of
predictors (but also works for dfs).}
}
\value{
A list with elements \code{ok} (whether the check passes) and
\code{bad_classes} (a named list of any bad classes, where the name is the name
of the column, and the value is the class).
}
\description{
Identify Non-Character Predictors
}
\keyword{internal}
