% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_output.R
\name{.finalize_embeddings}
\alias{.finalize_embeddings}
\title{Finalize Embeddings Output}
\usage{
.finalize_embeddings(processed_embeddings, token_map)
}
\arguments{
\item{processed_embeddings}{The output from \code{.process_embeddings_result}.}

\item{token_map}{The output from \code{.make_token_map}.}
}
\value{
A tidy version of the output embeddings, with one embedding vector
per row.
}
\description{
Takes the data frame output from \code{.process_embeddings_result}, joins on the
\code{token_map} information and filters out the padding tokens.
}
\keyword{internal}
