% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert_classification-predict.R
\name{.get_bert_classification_predict_function}
\alias{.get_bert_classification_predict_function}
\title{Choose the Proper Prediction Function for BERT Classification}
\usage{
.get_bert_classification_predict_function(type)
}
\arguments{
\item{type}{A single character. The type of predictions to generate. Valid
options are:
\itemize{
\item \code{"class"} for "hard" class predictions.
\item \code{"prob"} for class probabilities.
}}
}
\value{
The predict function.
}
\description{
Choose the Proper Prediction Function for BERT Classification
}
\keyword{internal}
