% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{.is_null_external_pointer}
\alias{.is_null_external_pointer}
\title{Check if a Pointer is NULL}
\usage{
.is_null_external_pointer(pointer)
}
\arguments{
\item{pointer}{A pointer to check.}
}
\value{
A logical indicating whether that pointer is NULL (and thus invalid).
}
\description{
This comes from https://stackoverflow.com/a/27350487/3297472 via
https://github.com/mlverse/tabnet/blob/main/R/hardhat.R#L494
}
\keyword{internal}
