% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_output.R
\name{.make_index_column}
\alias{.make_index_column}
\title{Make Index Column}
\usage{
.make_index_column(shape, which)
}
\arguments{
\item{shape}{Integer vector; the shape of the tensor to index.}

\item{which}{Integer scalar; which dimension to index.}
}
\value{
The index column as an integer vector.
}
\description{
Given an overall shape and a particular dimension, construct the index column
for that dimension.
}
\keyword{internal}
