% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert_classification-fit.R
\name{.mold_valid_formula}
\alias{.mold_valid_formula}
\title{Mold Validation Data}
\usage{
.mold_valid_formula(formula, valid_data, blueprint)
}
\arguments{
\item{formula}{A formula specifying the outcome term on the left-hand side,
and the predictor terms on the right-hand side.}

\item{valid_data}{When a \strong{formula} is used, \code{valid_data} can be:
\itemize{
\item A \strong{data frame} containing both the predictors and the outcome to be
used for model validation, in the same format as \code{data}.
\item A number between 0 and 1, representing the fraction of data to use for
model validation.
\item \code{NULL}, in which case no data will be used for model validation.
}}

\item{blueprint}{The blueprint generated by \code{\link[hardhat]{mold}} for
the training data.}
}
\value{
The processed validation data.
}
\description{
Mold Validation Data
}
\keyword{internal}
