% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert_classification-fit.R
\name{.mold_valid_xy}
\alias{.mold_valid_xy}
\title{Mold Validation Data}
\usage{
.mold_valid_xy(valid_x, valid_y, blueprint)
}
\arguments{
\item{valid_x}{Depending on the context:
\itemize{
\item A number between 0 and 1, representing the fraction of data to use for
model validation.
\item Predictors in the same format as \code{x}. These predictors will be used for
model validation.
\item \code{NULL}, in which case no data will be used for model validation.
}}

\item{valid_y}{When \code{valid_x} is a set of predictors, \code{valid_y} should be
outcomes in the same format as \code{y}.}

\item{blueprint}{The blueprint generated by \code{\link[hardhat]{mold}} for
the training data.}
}
\value{
The processed validation data.
}
\description{
Mold Validation Data
}
\keyword{internal}
