% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert_classification-fit.R
\name{.new_bert_classification}
\alias{.new_bert_classification}
\title{Create the Bert Classification Object}
\usage{
.new_bert_classification(luz_model, outcome_levels, n_tokens, blueprint)
}
\arguments{
\item{luz_model}{A model fit using \code{{luz}}.}

\item{outcome_levels}{The levels of the outcome factor.}

\item{n_tokens}{The number of tokens to which the predictors should be
tokenized.}

\item{blueprint}{The \code{{hardhat}} blueprint to reproduce the data.}
}
\value{
A \code{{hardhat}} \code{\link[hardhat]{new_model}}.
}
\description{
Create the Bert Classification Object
}
\keyword{internal}
