% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert_classification-predict.R
\name{.predict_bert_classification_bridge}
\alias{.predict_bert_classification_bridge}
\title{Prepare BERT Classification Data for Prediction}
\usage{
.predict_bert_classification_bridge(type, object, predictors)
}
\arguments{
\item{type}{A single character. The type of predictions to generate.
Valid options are:
\itemize{
\item \code{"class"} for "hard" class predictions.
\item \code{"prob"} for class probabilities.
}}

\item{object}{A \code{bert_classification} object.}

\item{predictors}{Forged predictor data.}
}
\value{
A tibble with output dependent on the type.
}
\description{
Prepare BERT Classification Data for Prediction
}
\keyword{internal}
