% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert_classification-predict.R
\name{.predict_bert_classification_class}
\alias{.predict_bert_classification_class}
\title{Get Class for BERT Classification}
\usage{
.predict_bert_classification_class(object, predictors)
}
\arguments{
\item{object}{A \code{bert_classification} object.}

\item{predictors}{Forged predictor data.}
}
\value{
A tibble with a factor column identifying the most probable outcome.
}
\description{
Get Class for BERT Classification
}
\keyword{internal}
