% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert_classification-predict.R
\name{.predict_bert_classification_prob}
\alias{.predict_bert_classification_prob}
\title{Get Probabilities for BERT Classification}
\usage{
.predict_bert_classification_prob(object, predictors)
}
\arguments{
\item{object}{A \code{bert_classification} object.}

\item{predictors}{Forged predictor data.}
}
\value{
A tibble of probabilities.
}
\description{
Get Probabilities for BERT Classification
}
\keyword{internal}
