% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert_classification-predict.R
\name{.predict_bert_classification_shared}
\alias{.predict_bert_classification_shared}
\title{Shared Prediction for BERT Classification}
\usage{
.predict_bert_classification_shared(object, predictors)
}
\arguments{
\item{object}{A \code{bert_classification} object.}

\item{predictors}{Forged predictor data.}
}
\value{
A torch tensor of softmax probabilities.
}
\description{
Shared Prediction for BERT Classification
}
\keyword{internal}
