% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert_regression-predict.R
\name{.predict_bert_regression}
\alias{.predict_bert_regression}
\title{Get Predictions for BERT Regression}
\usage{
.predict_bert_regression(object, predictors)
}
\arguments{
\item{object}{A \code{bert_regression} object.}

\item{predictors}{Forged predictor data.}
}
\value{
A tibble of numeric predictions
}
\description{
Get Predictions for BERT Regression
}
\keyword{internal}
