% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert_regression-predict.R
\name{.predict_bert_regression_bridge}
\alias{.predict_bert_regression_bridge}
\title{Prepare BERT Regression Data for Prediction}
\usage{
.predict_bert_regression_bridge(object, predictors)
}
\arguments{
\item{object}{A \code{bert_regression} object.}

\item{predictors}{Forged predictor data.}
}
\value{
A tibble with output dependent on the type.
}
\description{
Prepare BERT Regression Data for Prediction
}
\keyword{internal}
