% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{.serialize_luz_model}
\alias{.serialize_luz_model}
\title{Serialize a Luz Model}
\usage{
.serialize_luz_model(luz_model)
}
\arguments{
\item{luz_model}{A \code{luz_model} object to update.}
}
\value{
The \code{luz_model} with \code{serialized_state_dict} attached.
}
\description{
Parameters for luz models are held in pointers that become invalid if you
save and then load the model. To avoid issues with those pointers, we add a
\code{serialized_state_dict} parameter to the \code{luz_model} object, which holds safe
versions of the parameter values.
}
\keyword{internal}
