% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{.unserialize_luz_model}
\alias{.unserialize_luz_model}
\title{Restore a Serialized State Dictionary}
\usage{
.unserialize_luz_model(luz_model)
}
\arguments{
\item{luz_model}{The model to fix.}
}
\value{
The model, invisibly.
}
\description{
This function reloads the state dictionary in place.
}
\keyword{internal}
