% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{.validate_predictor_count}
\alias{.validate_predictor_count}
\title{Confirm the Number of Predictors}
\usage{
.validate_predictor_count(predictors, max_predictors = 2)
}
\arguments{
\item{predictors}{"Data-like" predictors, in this case always a tibble of
predictors (but also works for dfs).}

\item{max_predictors}{The maximum expected number of predictors.}
}
\value{
\code{predictors} (invisibly), or an error if there are too many columns.
}
\description{
Confirm the Number of Predictors
}
\keyword{internal}
