% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{.validate_predictors_are_character}
\alias{.validate_predictors_are_character}
\title{Confirm that Predictors are Character}
\usage{
.validate_predictors_are_character(predictors)
}
\arguments{
\item{predictors}{"Data-like" predictors, in this case always a tibble of
predictors (but also works for dfs).}
}
\value{
\code{predictors} (invisibly), or an error identifying bad columns.
}
\description{
Confirm that Predictors are Character
}
\keyword{internal}
