% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert_linear.R
\name{model_bert_linear}
\alias{model_bert_linear}
\title{Pretrained BERT Model with Linear Output}
\usage{
model_bert_linear(bert_type = "bert_tiny_uncased", output_dim = 1L)
}
\arguments{
\item{bert_type}{Character; which flavor of BERT to use. See
\code{\link[torchtransformers:available_berts]{available_berts()}} for known models.}

\item{output_dim}{Integer; the target number of output dimensions.}
}
\value{
A torch neural net model with pretrained BERT weights and a final
dense layer.
}
\description{
Construct a BERT model with pretrained weights, and add a final dense linear
layer to transform to a desired number of dimensions. Note that we only use
the CLS token output from the final layer of the BERT model. It is possible
to attach a classification head to BERT using other techniques, but here we
use this simple technique.
}
