% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip.R
\name{n_tokens}
\alias{n_tokens}
\title{Number of Tokens}
\usage{
n_tokens(range = c(1, 9), trans = scales::log2_trans())
}
\arguments{
\item{range}{A two-element integer vector with the smallest and largest
possible values. By default these values should be the powers of two to
try.}

\item{trans}{An optional transformation to apply. By default,
\code{\link[scales:log_trans]{scales::log2_trans()}} (meaning take the log2 of the original values,
resulting in small-number range values).}
}
\value{
A parameter that can be tuned with the \code{tune} package.
}
\description{
The number of tokens to use for tokenization of predictors.
}
\examples{
if (rlang::is_installed("dials")) {
  n_tokens()
}
}
