% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert_regression-predict.R
\name{predict.bert_regression}
\alias{predict.bert_regression}
\title{Predict from a \code{bert_regression} model.}
\usage{
\method{predict}{bert_regression}(object, new_data, ...)
}
\arguments{
\item{object}{A \code{bert_regression} object.}

\item{new_data}{A data frame or matrix of new character predictors. This data
is automatically tokenized to match the tokenization expected by the BERT
model.}

\item{...}{Not used, but required for extensibility.}
}
\value{
A tibble of predictions. The number of rows in the tibble is guaranteed to be
the same as the number of rows in \code{new_data}.
}
\description{
Predict from a \code{bert_regression} model.
}
