% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_output.R
\name{tidy_bert_output}
\alias{tidy_bert_output}
\title{Tidy the BERT Output}
\usage{
tidy_bert_output(bert_model_output, tokenized)
}
\arguments{
\item{bert_model_output}{The output from a BERT model.}

\item{tokenized}{The raw output from \code{torchtransformers::tokenize_bert} with \code{simplify = FALSE}.}
}
\value{
A list of data frames, one for the layer output embeddings and one
for the attention weights.
}
\description{
Given the output from a transformer model, construct tidy data frames for
the layer outputs and the attention weights.
}
