% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidybert-package.R
\docType{package}
\name{tidybert-package}
\alias{tidybert}
\alias{tidybert-package}
\title{tidybert: Tidy BERT-like Models}
\description{
Implements BERT NLP models with a consistent interface for fitting and creating predictions. The models are fully compatible with the tidymodels framework.
}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

Authors:
\itemize{
  \item Jonathan Bratt \email{jonathan.bratt@macmillan.com} (\href{https://orcid.org/0000-0003-2859-0076}{ORCID})
}

Other contributors:
\itemize{
  \item Bedford Freeman & Worth Pub Grp LLC DBA Macmillan Learning [copyright holder]
}

}
\keyword{internal}
