% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip.R
\name{tidybert_engine}
\alias{tidybert_engine}
\title{BERT models via tidybert}
\description{
The \code{tidybert} engine defines a \code{\link[=bert_classification]{bert_classification()}} or
\code{\link[=bert_regression]{bert_regression()}} model. The models are fit using
\code{\link[luz:fit.luz_module_generator]{luz::fit.luz_module_generator()}}.
}
\details{
The available parameters are \code{bert_type}, \code{n_tokens}, \code{valid_data}, \code{loss},
\code{optimizer}, \code{metrics}, \code{luz_opt_hparams}, and \code{...}. See
\code{\link[=bert_classification]{bert_classification()}} and \code{\link[=bert_regression]{bert_regression()}} for details.
}
\keyword{internal}
