% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert-fit.R
\name{bert}
\alias{bert}
\alias{bert.default}
\alias{bert.data.frame}
\alias{bert.matrix}
\alias{bert.formula}
\alias{bert.recipe}
\title{Fit a \code{bert}}
\usage{
bert(x, ...)

\method{bert}{default}(x, ...)

\method{bert}{data.frame}(x, y, ...)

\method{bert}{matrix}(x, y, ...)

\method{bert}{formula}(formula, data, ...)

\method{bert}{recipe}(x, data, ...)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
\item A \strong{matrix} of predictors.
\item A \strong{recipe} specifying a set of preprocessing steps
created from \code{\link[recipes:recipe]{recipes::recipe()}}.
}}

\item{...}{Not currently used, but required for extensibility.}

\item{y}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{y} is the outcome
specified as:
\itemize{
\item A \strong{data frame} with 1 numeric column.
\item A \strong{matrix} with 1 numeric column.
\item A numeric \strong{vector}.
}}

\item{formula}{A formula specifying the outcome terms on the left-hand side,
and the predictor terms on the right-hand side.}

\item{data}{When a \strong{recipe} or \strong{formula} is used, \code{data} is specified as:
\itemize{
\item A \strong{data frame} containing both the predictors and the outcome.
}}
}
\value{
A \code{bert} object.
}
\description{
\code{bert()} fits a model.
}
\examples{
predictors <- mtcars[, -1]
outcome <- mtcars[, 1]

# XY interface
mod <- bert(predictors, outcome)

# Formula interface
mod2 <- bert(mpg ~ ., mtcars)

# Recipes interface
library(recipes)
rec <- recipe(mpg ~ ., mtcars)
rec <- step_log(rec, disp)
mod3 <- bert(rec, mtcars)

}
