% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert_classification-fit.R
\name{bert_classification}
\alias{bert_classification}
\alias{bert_classification.default}
\alias{bert_classification.data.frame}
\alias{bert_classification.matrix}
\alias{bert_classification.formula}
\alias{bert_classification.recipe}
\title{Fit a BERT-style neural network}
\usage{
bert_classification(
  x,
  model_name = "bert_tiny_uncased",
  n_tokens = torchtransformers::config_bert(model_name, "max_tokens"),
  ...
)

\method{bert_classification}{default}(x, ...)

\method{bert_classification}{data.frame}(x, y, ...)

\method{bert_classification}{matrix}(x, y, ...)

\method{bert_classification}{formula}(formula, data, ...)

\method{bert_classification}{recipe}(x, data, ...)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of character predictors.
\item A \strong{matrix} of character predictors.
\item Note that a \strong{recipe} created from \code{\link[recipes:recipe]{recipes::recipe()}} will NOT
currently work.
}}

\item{...}{Not currently used, but required for extensibility.}

\item{y}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{y} is the outcome
specified as:
\itemize{
\item A \strong{data frame} with 1 factor column.
\item A \strong{matrix} with 1 factor column.
\item A factor \strong{vector}.
}}

\item{formula}{A formula specifying the outcome term on the left-hand side,
and the predictor terms on the right-hand side.}

\item{data}{When a \strong{formula} is used, \code{data} is specified as:
\itemize{
\item A \strong{data frame} containing both the predictors and the outcome. The
predictors should be character vectors. The outcome should be a factor,
or a character vector.
}}
}
\value{
A \code{bert_classification} object.
}
\description{
\code{bert_classification()} fits a classifier neural network in the style of
\href{https://github.com/google-research/bert}{BERT from Google Research}.
}
