% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert_linear.R
\name{model_bert_linear}
\alias{model_bert_linear}
\title{Pretrained BERT Model with Linear Output}
\usage{
model_bert_linear(model_name = "bert_tiny_uncased", output_dim = 1L)
}
\arguments{
\item{model_name}{Character; which flavor of BERT to use. See
\code{\link[torchtransformers]{available_berts}} for known models.}

\item{output_dim}{Integer; the target number of output dimensions.}
}
\value{
A torch neural net model with pretrained BERT weights and a final
dense layer.
}
\description{
Construct a BERT model with pretrained weights, and add a final dense linear
layer to transform to a desired number of dimensions.
}
